requirements.txt
pandas>=2.1
pyarrow>=16.0
openpyxl>=3.1
xlrd>=2.0
PyMuPDF>=1.24
python-docx>=1.1
striprtf>=0.0.26
beautifulsoup4>=4.12
Pillow>=10.0
pytesseract>=0.3.10
opencv-python>=4.10
tqdm>=4.66
ijson>=3.3
streamlit>=1.35
pytest>=8.2
